cut: fix unsigned-wrap subscript UB in f_cut
p[i - clen] with int i and size_t clen computed p + (SIZE_MAX - ...),
wrapping the address space (UB). Use a signed subscript.

+++ src.freebsd/coreutils/cut/cut.c
@@ -450,7 +450,7 @@ f_cut(FILE *fp, const char *fname)
 				}
 				if (*pos)
 					for (i = 0; i < (int)clen; i++)
-						putchar(p[i - clen]);
+						putchar(p[i - (int)clen]);
 			}
 			if (ch == '\n')
 				break;
